+++ PSD_builder/src/main/kotlin/com/wip/psdbuilder/PSDBuilderPlugin.kt
@@ -43,7 +43,7 @@ data class PsdPayload(
 @PluginInfo(
     id = "com.wip.psdbuilder",
     name = "PSD Builder",
-    version = "2.0.2",
+    version = "2.1.0",
     description = "A plugin that builds layered PSD files using PSD_builder.exe."
 )
 class PSDBuilderPlugin {